requirements.txt
# Python 3.10 compatible versions
# Optimized for Apple Silicon M4 Pro with Metal acceleration
# Fixed compatibility issues with sentence-transformers and huggingface-hub

# Core dependencies - Metal/MPS compatible
torch>=2.0.0,<2.5.0  # MPS support for Apple Silicon
transformers>=4.30.0,<5.0.0
datasets>=2.14.0,<3.0.0
sentence-transformers>=2.7.0,<3.0.0  # Fixed: Compatible with newer huggingface_hub
huggingface-hub>=0.19.0,<1.0.0  # Fixed: Compatible version (no cached_download)

# Retrieval & Indexing - Apple Silicon optimized
faiss-cpu>=1.7.4,<2.0.0  # Optimized for ARM64
pyserini>=0.21.0,<1.0.0
rank-bm25>=0.2.2

# Data processing
pandas>=2.0.0,<2.3.0
numpy>=1.24.0,<2.0.0  # Stable for Python 3.10
tqdm>=4.65.0
nltk>=3.8.1,<4.0.0
spacy>=3.5.0,<4.0.0

# Evaluation
scikit-learn>=1.3.0,<1.6.0
scipy>=1.11.0,<1.15.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
wandb>=0.15.0,<1.0.0  # Optional: for experiment tracking
jupyter>=1.0.0
ipython>=8.12.0  # Better compatibility with Python 3.10
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# Apple Silicon specific optimizations
accelerate>=0.20.0  # For efficient model loading on Metal
# Optional: for ColBERT if you want to use it later
# colbert-ai>=0.2.0
